+++ R/object-package.R
@@ -1,25 +1,24 @@
-package_seealso <- function(desc) {
-  itemize("Useful links:", package_seealso_urls(desc))
+package_seealso <- function(URL, BugReports) {
+  itemize("Useful links:", package_seealso_urls(URL, BugReports))
 }
-package_seealso_urls <- function(desc) {
-  if (!is.null(desc$URL)) {
-    links <- paste0("\\url{", strsplit(desc$URL, ",\\s+")[[1]], "}")
+package_seealso_urls <- function(URL = NULL, BugReports = NULL) {
+  if (!is.null(URL)) {
+    links <- paste0("\\url{", strsplit(URL, ",\\s+")[[1]], "}")
     links <- gsub("\\url\\{https://doi.org/", "\\doi{", links)
   } else {
     links <- character()
   }
-  if (!is.null(desc$BugReports)) {
-    links <- c(links, paste0("Report bugs at \\url{", desc$BugReports, "}"))
+  if (!is.null(BugReports)) {
+    links <- c(links, paste0("Report bugs at \\url{", BugReports, "}"))
   }
 
   links
 }
 
-
-package_authors <- function(desc) {
-  authors <- tryCatch(eval(parse(text = desc$`Authors@R` %||% "")),
+package_authors <- function(authors) {
+  authors <- tryCatch(eval(parse(text = authors %||% "")),
     error = function(e) {
-      warning(e)
+      cli::cli_warn("Failed to evaluate Authors@R.", parent = e)
       NULL
     }
   )
@@ -42,7 +41,7 @@ package_authors <- function(desc) {
 
 author_desc <- function(x) {
   if (inherits(x, "person")) {
-    stop("person class must be stripped", call. = FALSE)
+    cli::cli_abort("Person class must be stripped", .internal = FALSE)
   }
 
   desc <- paste0(x$given, collapse = " ")