+++ tests/testthat/helper-test.R
@@ -16,9 +16,14 @@ expect_parse_failure <- function(code)  {
   (expect_warning(expect_null(code)))
 }
 
-local_package_copy <- function(path, env = caller_env()) {
+local_package_copy <- function(path, env = caller_env(), set_version = TRUE) {
   temp_path <- withr::local_tempdir(.local_envir = env)
 
   file.copy(path, temp_path, recursive = TRUE)
   withr::local_dir(file.path(temp_path, basename(path)), .local_envir = env)
+  if (set_version) {
+    desc::desc_set(RoxygenNote = as.character(packageVersion("roxygen2")))
+  }
+
+  invisible()
 }