+++ tests/testthat/test-topics.R
@@ -1,0 +1,7 @@
+test_that("missing title generates useful message", {
+  block <- "
+    #' @name foo
+    NULL
+  "
+  expect_snapshot(roc_proc_text(rd_roclet(), block))
+})