+++ tests/testthat/testNonASCII/R/a.r
@@ -2,5 +2,4 @@
 #'
 #' @note 我爱中文。
 printChineseMsg <- function() {
-  message("我是UTF8的中文字符。")
 }